% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitloop.r
\name{logitloop}
\alias{logitloop}
\title{logitloop
this function makes univariate logit analyses}
\usage{
logitloop(expression,outcome="os.status")
}
\description{
logitloop
this function makes univariate logit analyses
}
\examples{
library(logitloop)
library(dplyr)
data(kras)
kras\%>\%select(-os.months)->expression
df<-logitloop(expression,outcome="os.status")
df
}
