% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoef.r
\name{plotcoef}
\alias{plotcoef}
\title{plotcoef
this function plot the results of logit analyses}
\usage{
plotcoef(df,nb=11,title="")
}
\description{
plotcoef
this function plot the results of logit analyses
}
\examples{
library(logitloop)
library(dplyr)
data(kras)
kras\%>\%select(-os.months)->expression
df<-logitloop(expression,outcome="os.status")
df
plotcoef(df,nb=11,title="")
}
